requirements.txt
Flask==2.0.1
Flask-Toastr
Flask-Admin==1.5.8
Flask-Login==0.5.0
Flask-SocketIO==5.1.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
Jinja2==3.0.1
Paginator==0.5.1
passlib==1.7.4
py==1.10.0
py-paginator==1.0.2
PyMySQL==1.0.2
PyNaCl==1.4.0
pyparsing==2.4.7
pyRFC3339==1.1
pysha3==1.0.2
pytest==6.2.5
python-socketio==5.4.0
PyYAML==5.3.1
requests==2.23.0
requests-unixsocket==0.2.0
SecretStorage==3.1.2
simplejson==3.17.0
SQLAlchemy==1.4.22
virtualenv==20.7.1
Werkzeug==2.0.1
WTForms==2.3.3